Moving NotesView
